requirements.txt
streamlit
pdfminer.six
sentence-transformers
scikit-learn
groq
matplotlib
python-dotenv
pymupdf
